asm/fp: interface refactor
Refactors the interface into a "two phase" initialization. A field can
be constructed as an object first, and then an avo context can be
provided later (creating a Builder). Now the avo context is a struct
field instead of provided as an argument to every function. The
motivation here is it will allow addtional state to be stored alongside
the context, for example initialized global data sections. This will be
useful for the implementation of Montgomery fields.

Updates #47