gen/curve: scalar multiplication (#80)
Provides the bulk of an implementation of scalar multiplication for short Weierstrass curves. Follows Algorithm 1 in "Selecting Elliptic Curves for Cryptography: An Efficiency and Security Analysis" by Bos et al.

Note this is not quite complete. The final add needs to use a complete addition formula, and ensure that the zero scalar is handled correctly.

Updates #67 #55